requirements.txt
discord
python-dotenv
PyNaCl